Split Explorer.swift into multiple files and more - Rename LngLat to Coordinate - Add S2Cell.neighboredCells(in:)
